+++ Assets/Prefabs/Enemies/bombardment/BombardmentCrosshair.prefab
@@ -98,7 +98,8 @@ GameObject:
   - component: {fileID: 8483520523054886789}
   - component: {fileID: 8732174585770467874}
   - component: {fileID: 540275300880896012}
-  - component: {fileID: 223185426903044165}
+  - component: {fileID: 4663268476541556188}
+  - component: {fileID: 6490354721274685463}
   m_Layer: 7
   m_Name: BombardmentCrosshair
   m_TagString: Crosshair
@@ -119,7 +120,6 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 752439470486415171}
   - {fileID: 1871905132550475885}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -186,7 +186,7 @@ CircleCollider2D:
   m_CompositeOrder: 0
   m_Offset: {x: 0, y: 0}
   m_Radius: 0.48
---- !u!114 &223185426903044165
+--- !u!114 &4663268476541556188
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -195,65 +195,28 @@ MonoBehaviour:
   m_GameObject: {fileID: 1725758782596052211}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 77cf0280af908d942ac0963ddbde67c7, type: 3}
+  m_Script: {fileID: 11500000, guid: 6e6cbf9f5defc2541a8fd7fe7d82a313, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hurtbox: {fileID: 3403728834270932085}
-  crosshairMoveSpeed: 0
+  hurtboxSizeReference: {fileID: 6490354721274685463}
   damage: 1
-  fireRate: 0
-  hurtboxDuration: 0.1
+  crosshairMoveSpeed: 0
+  lifeTime: 2
   rb: {fileID: 8732174585770467874}
-  pauseTimeUntilFire: 1
-  pauseTimeAfterFire: 0.2
-  angleRange: 0
---- !u!1 &3403728834270932085
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 752439470486415171}
-  - component: {fileID: 8460390129649112995}
-  - component: {fileID: 1300480048583582568}
-  m_Layer: 0
-  m_Name: Hurtbox
-  m_TagString: Crosshair
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!4 &752439470486415171
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3403728834270932085}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 8483520523054886789}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!58 &8460390129649112995
+--- !u!58 &6490354721274685463
 CircleCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3403728834270932085}
+  m_GameObject: {fileID: 1725758782596052211}
   m_Enabled: 1
   serializedVersion: 3
   m_Density: 1
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
-    m_Bits: 0
+    m_Bits: 256
   m_ExcludeLayers:
     serializedVersion: 2
     m_Bits: 0
@@ -276,15 +239,3 @@ CircleCollider2D:
   m_CompositeOrder: 0
   m_Offset: {x: 0, y: 0}
   m_Radius: 0.35
---- !u!114 &1300480048583582568
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3403728834270932085}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f4c1e9fd453520b4e84455a5f309d714, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
