--- Assets/Scripts/Enemies/BasicEnemy.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7852888ba22165d4f993c8dd6cbfaf2f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
