--- Assets/Scripts/Enemies/CrosshairHurtbox.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f4c1e9fd453520b4e84455a5f309d714
