--- Assets/Scripts/Enemies/InnocentEnemy.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e9fa7c0918bc52041b3cbf645a78d00c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
